+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Extension "Geodata" for Contao Open Source CMS
 1.0.0-rc6 - 2023-07-17
 ---
 - UPDATED : map items can now have multiple categories
+- UPDATED : filters in locations' list
 
 1.0.0-rc5 - 2023-02-06
 ---